extract spectrum legend creator to own module also, corrected the indentation (hopefully for the last time) removed commented out code
