+++ index.html
@@ -11,7 +11,6 @@
       <div class="plot-container"></div>
       <p class="side-table" id="demo3"></p>
     </div>
-    <!-- <p class="p9"><b>Circle</b><input class="input1" type="checkbox" id="cbox4" value="fourth_checkbox"></input><input class = "input2" type="text" id="drawxy" placeholder="Select 1 pt after check"> </input> <button class = "input2" onclick="handleClick4()" value="Draw" >Draw</button></p> -->
     <div class="plot-options-container">
       <div class="options-container zoom-container">
         <div class="enable-zoom">
